requirements.txt
# K-Means Image Compression Flask App
# Production dependencies

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.1

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Production Server
gunicorn>=21.0.0

# Environment Variables
python-dotenv>=1.0.0
